try to add ControllerAdvice

+++ store-service/src/main/resources/db/migration/V2__Images.sql
@@ -1,6 +1,6 @@
 create table if not exists store.images
 (
-    id           bigint primary key,
+    id           serial primary key,
     file_name    varchar not null unique,
     o_file_name  varchar not null,
     size         bigint,
